Minimally functional conditional statements
The implementation is rather constraining because of challenges related
to newlines as statement delimiters. At present, it is not known how to
mix newline statement delimiters with newlines in conditional statements
and maintain an unambiguous grammar. This leads to the requirement that
unbraced conditional blocks must be on a single line, and braced
conditional blocks must have collapsed opening braces.